requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.6
django-cors-headers==3.14.0
django-storages==1.13.2
djangochannelsrestframework==1.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
urllib3==1.26.16
whitenoise==6.4.0
